New reusable mock for bottomSheetIntegration
