+++ app/screens/BottomSheetScreen/__test__/BottomSheetScreen.test.tsx
@@ -1,27 +1,10 @@
+import { mockBottomSheetModal } from '@app/testing/mockBottomSheetModal';
 import { fireEvent, render } from '@testing-library/react-native';
 import React from 'react';
 
 import { BottomSheetScreen } from '../BottomSheetScreen';
 
-const mockOpenModal = jest.fn();
-const mockCloseModal = jest.fn();
-
-jest.mock('@app/modal/BottomSheetModal', () => {
-  const { SNAP_POINTS_DYNAMIC, SNAP_POINTS_MEDIUM_FULL_HEIGHT } = jest.requireActual(
-    '@app/modal/BottomSheetModal',
-  );
-  return {
-    SNAP_POINTS_DYNAMIC,
-    SNAP_POINTS_MEDIUM_FULL_HEIGHT,
-    bottomSheetModalRef: {
-      _test: 'hello',
-      current: {
-        open: (params: any) => mockOpenModal(params),
-        close: () => mockCloseModal(),
-      },
-    },
-  };
-});
+const { mockOpenModal } = mockBottomSheetModal();
 
 describe('BottomSheetScreen', () => {
   it('should open linkedInModal with the right parameters', () => {
